+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/KryptonColorTable2007.cs
@@ -499,7 +499,7 @@ public class KryptonColorTable2007 : KryptonColorTable
     /// <summary>
     /// Gets the text color used on the tool strip.
     /// </summary>
-    public override Color ToolStripText => Colors[(int)SchemeBaseColors.TextButtonNormal];
+    public override Color ToolStripText => Colors.ResolveToolStripText(SchemeBaseColors.TextButtonNormal);
 
     #endregion