+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/KryptonColorTableMicrosoft365.cs
@@ -506,7 +506,7 @@ public class KryptonColorTableMicrosoft365 : KryptonColorTable
     /// <summary>
     /// Gets the text color used on the tool strip.
     /// </summary>
-    public override Color ToolStripText => _colors[(int)SchemeBaseColors.StatusStripText];
+    public override Color ToolStripText => _colors.ResolveToolStripText(SchemeBaseColors.StatusStripText);
 
     #endregion