+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/KryptonVisualStudio2010With2007ColorTable.cs
@@ -497,7 +497,7 @@ public class KryptonVisualStudio2010With2007ColorTable : KryptonColorTable
     /// <summary>
     /// Gets the text color used on the tool strip.
     /// </summary>
-    public override Color ToolStripText => Colors[(int)SchemeBaseColors.TextButtonNormal];
+    public override Color ToolStripText => Colors.ResolveToolStripText(SchemeBaseColors.TextButtonNormal);
 
     #endregion