+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/KryptonVisualStudio2010With2013ColorTable.cs
@@ -504,7 +504,7 @@ public class KryptonVisualStudio2010With2013ColorTable : KryptonColorTable
     /// <summary>
     /// Gets the text color used on the tool strip.
     /// </summary>
-    public override Color ToolStripText => _colors[(int)SchemeBaseColors.StatusStripText];
+    public override Color ToolStripText => _colors.ResolveToolStripText(SchemeBaseColors.StatusStripText);
 
     #endregion